⚡️ Improve performances
