+++ nuxt.config.ts
@@ -25,7 +25,7 @@ export default defineNuxtConfig({
   css: ['@unocss/reset/tailwind.css', '/assets/css/main.css'],
   image: {
     dir: 'assets',
-    quality: 95,
+    quality: 90,
   },
   shiki: {
     bundledLangs: ['json', 'javascript'],
@@ -51,39 +51,7 @@ export default defineNuxtConfig({
   },
   nitro: {
     prerender: {
-      routes: [
-        '/',
-        '/_ipx/f_gif&q_95&s_200x43/ufo.gif',
-        '/_ipx/f_gif&q_95&s_400x86/ufo.gif',
-        '/_ipx/f_gif&q_95&s_36x36/earth.gif',
-        '/_ipx/f_gif&q_95&s_72x72/earth.gif',
-        '/_ipx/f_gif&q_95&s_96x66/construction.gif',
-        '/_ipx/f_gif&q_95&s_192x132/construction.gif',
-        '/_ipx/f_gif&q_95&s_50x82/cmcdconstruction.gif',
-        '/_ipx/f_gif&q_95&s_100x164/cmcdconstruction.gif',
-        '/_ipx/f_gif&q_95&s_90x85/construction2.gif',
-        '/_ipx/f_gif&q_95&s_180x170/construction2.gif',
-        '/_ipx/f_gif&q_95&s_78x53/building.gif',
-        '/_ipx/f_gif&q_95&s_156x106/building.gif',
-        '/_ipx/f_gif&q_95&s_90x200/alien.gif',
-        '/_ipx/f_gif&q_95&s_180x400/alien.gif',
-        '/_ipx/f_gif&q_95&s_416x86/text.gif',
-        '/_ipx/f_gif&q_95&s_208x43/text.gif',
-        '/_ipx/f_gif&q_95&s_60x60/mail.gif',
-        '/_ipx/f_gif&q_95&s_120x120/mail.gif',
-        '/_ipx/f_webp&q_95&s_88x31/valid-html40.png',
-        '/_ipx/f_webp&q_95&s_176x62/valid-html40.png',
-        '/_ipx/f_webp&q_95&s_88x31/valid-css2.png',
-        '/_ipx/f_webp&q_95&s_176x62/valid-css2.png',
-        '/_ipx/f_gif&q_95&s_141x52/contact.gif',
-        '/_ipx/f_gif&q_95&s_282x104/contact.gif',
-        '/_ipx/f_gif&q_95&s_107x35/email.gif',
-        '/_ipx/f_gif&q_95&s_214x70/email.gif',
-        '/_ipx/f_gif&q_95&s_575x10/bluline.gif',
-        '/_ipx/f_gif&q_95&s_1150x20/bluline.gif',
-        '/_ipx/q_95&f_webp&s_160x160/avatar.jpg',
-        '/_ipx/q_95&f_gif&s_80x80/coding.gif',
-      ],
+      routes: ['/'],
     },
   },
 })
